ai/requirements.txt
Flask
python-dotenv
google-genai
pydantic
flask-cors
gunicorn